+++ default.nix
@@ -1,12 +1,12 @@
 { system ? builtins.currentSystem
 , profiling ? false
-, iosSdkVersion ? "13.2"
+, iosSdkVersion ? "16.1"
 , config ? {}
 , terms ? { # Accepted terms, conditions, and licenses
     security.acme.acceptTerms = false;
   }
 , reflex-platform-func ? import ./dep/reflex-platform
-, useGHC810 ? false #true if one wants to use ghc 8.10.7
+, useGHC810 ? true # false if one wants to use ghc 8.6.5
 }:
 let
   reflex-platform = getReflexPlatform { inherit system; };
@@ -90,7 +90,8 @@ in rec {
       ${if optimizationLevel == null then ''
         ln -s "$dir/all.unminified.js" "$dir/all.js"
       '' else ''
-        '${pkgs.closurecompiler}/bin/closure-compiler' ${if externs == null then "" else "--externs '${externs}'"} --externs '${reflex-platform.ghcjsExternsJs}' -O '${optimizationLevel}' --jscomp_warning=checkVars --warning_level=QUIET --create_source_map="$dir/all.js.map" --source_map_format=V3 --js_output_file="$dir/all.js" "$dir/all.unminified.js"
+        # NOTE: "--error_format JSON" avoids closurecompiler crashes when trying to report errors.
+        '${pkgs.closurecompiler}/bin/closure-compiler' --error_format JSON ${if externs == null then "" else "--externs '${externs}'"} --externs '${reflex-platform.ghcjsExternsJs}' -O '${optimizationLevel}' --jscomp_warning=checkVars --warning_level=QUIET --create_source_map="$dir/all.js.map" --source_map_format=V3 --js_output_file="$dir/all.js" "$dir/all.unminified.js"
         echo '//# sourceMappingURL=all.js.map' >> "$dir/all.js"
       ''}
     done
@@ -332,10 +333,7 @@ in rec {
 
                 shellToolOverrides = lib.composeExtensions
                   self.userSettings.shellToolOverrides
-                  (if self.userSettings.__withGhcide
-                    then (import ./haskell-overlays/ghcide.nix)
-                    else (_: _: {})
-                  );
+                  (_: _: {});
 
                 project = reflexPlatformProject ({...}: self.projectConfig);
                 projectConfig = {
@@ -346,6 +344,10 @@ in rec {
                   shells = {
                     ${if self.userSettings.android == null && self.userSettings.ios == null then null else "ghcSavedSplices"} =
                       lib.filter (x: lib.hasAttr x self.combinedPackages) self.shells-ghcSavedSplices;
+
+                    ${if self.userSettings.ios == null then null else "ghcIosAarch64"} =
+                      lib.filter (x: lib.hasAttr x self.combinedPackages) self.shells-ghcSavedSplices;
+
                     ghc = lib.filter (x: lib.hasAttr x self.combinedPackages) self.shells-ghc;
                     ghcjs = lib.filter (x: lib.hasAttr x self.combinedPackages) self.shells-ghcjs;
                   };
@@ -403,6 +405,7 @@ in rec {
               `finally` writeProfilingData (profFileName ++ ".rprof")
         '';
       in nixpkgs.runCommand "ob-run" {
+        nativeBuildInputs = if system == "aarch64-darwin" then [ nixpkgs.darwin.cctools nixpkgs.darwin.signingUtils ] else [];
         buildInputs = [ (profiled.ghc.ghcWithPackages (p: [p.backend p.frontend])) ];
       } ''
         cp ${exeSource} ob-run.hs